+++ front/src/modules/companies/components/CompanyBoard/NewCompanyBoardCard.tsx
