+++ front/src/modules/companies/components/CompanyBoard/__stories__/Board.stories.tsx
