+++ front/src/modules/companies/components/CompanyBoard/__stories__/CompanyBoardCard.stories.tsx
@@ -1,11 +1,11 @@
 import { useState } from 'react';
 import { Meta, StoryObj } from '@storybook/react';
 
+import { Company } from '~/generated/graphql';
+import { mockedCompaniesData } from '~/testing/mock-data/companies';
+import { mockedPipelineProgressData } from '~/testing/mock-data/pipeline-progress';
 import { getRenderWrapperForComponent } from '~/testing/renderWrappers';
 
-import { Company } from '../../../../generated/graphql';
-import { mockedCompaniesData } from '../../../../testing/mock-data/companies';
-import { mockedPipelineProgressData } from '../../../../testing/mock-data/pipeline-progress';
 import { CompanyBoardCard } from '../CompanyBoardCard';
 
 const meta: Meta<typeof CompanyBoardCard> = {
