+++ front/src/modules/companies/components/CompanyBoard/__stories__/mock-data.ts
